Updating the Client class.
