+++ On The Map/Scenes/TabBarScene/TabBarInteractor.swift
@@ -15,6 +15,7 @@ protocol TabBarInteracting: AnyObject {
 }
 
 class TabBarInteractor {
+    private lazy var apiService = APIService()
     private let presenter: TabBarPresenting
     
     init(presenter: TabBarPresenting) {
@@ -25,7 +26,7 @@ class TabBarInteractor {
 extension TabBarInteractor: TabBarInteracting {
     func getAllLocations(completion: @escaping (Bool?) -> Void) {
         presenter.startLoading()
-        Client.getAllLocations { [weak self] error in
+        apiService.getAllLocations { [weak self] error in
             self?.presenter.stopLoading()
             if error == nil {
                 completion(true)
@@ -43,7 +44,7 @@ extension TabBarInteractor: TabBarInteracting {
     }
     
     func logout() {
-        Client.logout { [weak self] success, error in
+        apiService.logout { [weak self] success, error in
             if success {
                 self?.presenter.logout()
             } else {
